+++ rand_distr/src/unit_disc.rs
@@ -24,6 +24,7 @@ use rand::Rng;
 /// println!("{:?} is from the unit Disc.", v)
 /// ```
 #[derive(Clone, Copy, Debug)]
+#[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
 pub struct UnitDisc;
 
 impl<F: Float + SampleUniform> Distribution<[F; 2]> for UnitDisc {