+++ rand_distr/src/unit_sphere.rs
@@ -28,6 +28,7 @@ use rand::Rng;
 ///       Sphere.*](https://doi.org/10.1214/aoms/1177692644)
 ///       Ann. Math. Statist. 43, no. 2, 645--646.
 #[derive(Clone, Copy, Debug)]
+#[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
 pub struct UnitSphere;
 
 impl<F: Float + SampleUniform> Distribution<[F; 3]> for UnitSphere {